+++ bin/dns-scanner.js
@@ -3,14 +3,15 @@
 const DNSScanner = require('../src');
 const fs = require('fs');
 const path = require('path');
-const ProgressBar = require('progress');
 const program = require('commander');
-
-const start = new Date().getTime();
+const chalk = require('chalk');
+const logUpdate = require('log-update');
 
 program
     .usage('[options] <domains ...>')
-    .option('-f, --file [value]', 'dictionary file')
+    .option('-f, --file [value]', 'Dictionary file')
+    .option('-r, --recursive', 'Recursively scan sub-domains')
+    .option('-v, --verbose', 'Show scanning details info')
     .parse(process.argv);
 
 (async () => {
@@ -18,7 +19,13 @@ program
         return program.help();
     }
 
-    for (const domain of program.args) {
+    let domains = program.args;
+
+    const start = new Date().getTime();
+
+    do {
+        const domain = domains.shift();
+
         await new Promise((resolve, reject) => {
             let dictionaryFile = program.file || path.join(__dirname, '..', 'share', 'subdomains-10000.txt');
             const prefixes = fs
@@ -35,23 +42,36 @@ program
 
             scanner.start();
 
-            console.log(`[*] Scanning ${target} for A records:\n`);
+            const total = prefixes.length;
+            let current = 0;
+            let prevProgress;
 
-            const bar = new ProgressBar('Scanning... [:bar] :percent :etas', { total: prefixes.length, clear: true });
+            logUpdate.clear();
+
             scanner.on('progress', () => {
-                bar.tick();
+                current += 1;
+                let progress = Math.round((current / total) * 100);
+                if (prevProgress !== progress) {
+                    logUpdate(`::: Scanning ${target}: ${progress}% :::`);
+                    prevProgress = progress;
+                }
             });
 
             scanner.on('item', ({ address, ips }) => {
                 if (ips) {
-                    bar.interrupt(`${address} - ${ips.toString()}`);
+                    logUpdate.clear();
+                    console.log(`${chalk.bold(address)} - ${ips.toString()}`);
                 }
             });
 
             scanner.on('done', res => {
-                const end = new Date().getTime();
-                const duration = ((end - start) / 1000).toFixed(2);
-                console.log(`\nScanning of ${target} done! Found ${res.length} domains in ${duration} secs.\n`);
+                if (program.recursive) {
+                    for (const item of res) {
+                        domains.push(item.address);
+                    }
+                }
+
+                logUpdate.clear();
 
                 resolve();
             });
@@ -60,5 +80,12 @@ program
                 reject(err);
             });
         });
+    } while (domains.length > 0);
+
+    const end = new Date().getTime();
+    const duration = ((end - start) / 1000).toFixed(2);
+    if (program.verbose) {
+        logUpdate.clear();
+        console.log(`\n[✓] Scanning done in ${duration} seconds!`);
     }
 })();