+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-0.0.1 / 2018-08-03
+1.0.0 / 2018-08-03
 ==================
 
   * Initial release
+
+1.1.0 / 2018-08-27
+==================
+
+  * Recursively scan sub-domains
+  * Replace progressbar with raw progress message